Improve the write of UDP packets
Sometimes the TCP/IP stack can send misordered UDP packets. Since the transport protocol is UDP (instead of RTP), some TS packets may be lost on the client side. This problem occurs when two UDP packets are received in a different order. We have detected that in some circumstances the swapping of packets occurs on sending, and not on receiving. So this tries to fix it.